+++ services/budgets/BudgetService.ts
@@ -3,8 +3,6 @@ import {
   attributesToPatch,
   expenseInfo,
   patchableAttributes,
-  readBudgetComplete,
-  readBudgetInfo,
 } from '../../core/@types/budget/types';
 import { Budget } from '../../core/Budget';
 import BudgetRepository from '../../core/interfaces/budget/BudgetRepository';
